Allow Rails applications to override default views
The first time a template is rendered, Vanity checks to see if the path
'app/views/vanity' exists, and if so will load templates from that
location. Otherwise templates are loaded from the gem as normal.

This commit also adds a Rails generator to put copies of the original
templates into 'app/views/vanity'.