Refactor connection to allow specs to be built elsewhere
Now that more adapter tests are being added, it's useful to be able to
instantiate an adapter without going via the Connection class.

However, most of the test setup provides connection specs as strings,
requiring an extra parsing step that's currently baked into Connection, making
it hard for others to use.

This extracts the spec parsing into a Specification class for external use,
without altering the behaviour of Connection.