Use exposed specification to set up adapters
While most of the adapters work with an empty options hash, the mongo adapter
requires more configuration to instantiate successfully.

I've split out the test config in test_helper.rb to allow direct access to the
different configurations, and used the previously-introduced
Connection::Specification class to parse these strings into options that can
be passed to the adapters.