Silence deprecation warnings from legacy connecting test
This test involves resetting Vanity, which sets a new logger at the default
level. This means that the legacy configuration method it supports prints a
deprecation to the test output.

We can capture the IO and assert the presence of a deprecation message
instead using minitest's #capture_io, setting a new logger within the captured
block.