Deprecate #ab_seen with Alternative instance
Currently all methods in the storage adapter interface that accept an
`alternative` argument accept it in the form of an alternative ID. The
exception to this is `#ab_seen`, which accepts an Alternative instance.

Since third-party code might depend on this, this change at first deprecates
the old usage while allowing both calling styles. The old usage can then be
removed in a future version.