Explicitly enumerate routes in dummy test app
In the Rails 5.1 appraisal group, the following warning is generated:

    DEPRECATION WARNING: Using a dynamic :controller segment in a route
    is deprecated and will be removed in Rails 5.2. (called from block
    in <top (required)> at
    /Users/simon/dev/clones/vanity/test/dummy/config/route
    s.rb:57)

This is because of the wildcard route in the dummy test app:

    match ':controller(/:action(/:id(.:format)))', :via => [:get, :post]

To avoid this message, this change explicitly enumerates the routes used
by the dummy app. This is a mild hassle, but could be reduced in future
if vanity were turned into a rails engine, meaning most of the routes
would be provided directly by the gem.