Update redises for Fixnum deprecation
When running the tests in ruby 2.5, a lot of warnings are generated
about ::Fixnum being deprecated, caused by the redis gem.

This updates the redis gems in our various gemfiles to avoid the
deprecation message.

Because redis-rb has gained a major version, I've split the changes:

* Main and rails51 gemfiles: latest redis-rb
* rails41 gemfiles: minimum supported version containing the deprecation
  fix (3.3.3)

To achieve this I've locked the rails41 gemfiles at redis-rb 3.3.3
(instead of using >=), so it's clear we want them to stay at the minimum
supported version possible.

I've allowed the redis-namespace versions to jump to the latest version
in all gemfiles since all gemfiles were already on a version well past
the specified minimum.